refactor: restructure components to make more sense
Make a new "Board" component and make sure the "Tile" component represents a tile.

re #8
